# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ambulance & Emergency Warehouse Management System API — Damascus
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions/out:
    post:
      operationId: createOutTransaction
      tags: [transactions]
      summary: Record an outbound transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OutTransactionInput"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
...
  /alerts/{id}/read:
    post:
      operationId: markAlertRead
      tags: [alerts]
      summary: Mark a single alert as read for the current user
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /alerts/{id}/resolve:
    post:
      operationId: resolveAlert
      tags: [alerts]
      summary: Manually resolve an alert
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "404":
          description: Alert not found or already resolved
...
components:
  schemas:
...
    Transaction:
      type: object
      required: [id, type, itemType, documentNumber, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [in, out, init, adjust, custody_out, custody_return, damage, central_return]
        itemType:
          type: string
          enum: [item, equipment]
        itemId:
          type: ["integer", "null"]
        itemName:
          type: ["string", "null"]
        itemUnit:
          type: ["string", "null"]
        equipmentId:
          type: ["integer", "null"]
        equipmentName:
          type: ["string", "null"]
        quantity:
          type: ["integer", "null"]
        recipientId:
          type: ["integer", "null"]
        recipientName:
          type: ["string", "null"]
        recipientPerson:
          type: ["string", "null"]
        exitReasonId:
          type: ["integer", "null"]
        exitReason:
          type: ["string", "null"]
        supplier:
          type: ["string", "null"]
        batchNumber:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
        supplySource:
          type: ["string", "null"]
          enum: [central_warehouses, null]
        documentDate:
          type: ["string", "null"]
        deliveryNoteNumber:
          type: ["string", "null"]
        deliveryNoteDate:
          type: ["string", "null"]
        internalDeliveryNoteNumber:
          type: ["string", "null"]
        internalDeliveryNoteDate:
          type: ["string", "null"]
        deliveryDestination:
          type: ["string", "null"]
          enum: [administrative_building, health_facility, ambulance_point, null]
        custodyHolderName:
          type: ["string", "null"]
        custodyNoteNumber:
          type: ["string", "null"]
        custodyDate:
          type: ["string", "null"]
        custodyLocation:
          type: ["string", "null"]
        returnCondition:
          type: ["string", "null"]
        reason:
          type: ["string", "null"]
        documentNumber:
          type: string
        details:
          type: ["object", "null"]
          description: Structured movement details (stock before/after, delta, equipment snapshots) — used by adjustment vouchers
        notes:
          type: ["string", "null"]
        createdByName:
          type: ["string", "null"]
        createdAt:
          type: string
...
    OutTransactionInput:
      type: object
      required: [itemType, itemId, quantity, recipientId, exitReasonId, internalDeliveryNoteNumber, internalDeliveryNoteDate, deliveryDestination]
      properties:
        itemType:
          type: string
          enum: [item]
        itemId:
          type: integer
        equipmentId:
          type: ["integer", "null"]
        quantity:
          type: integer
          minimum: 1
        recipientId:
          type: integer
        recipientPerson:
          type: ["string", "null"]
        exitReasonId:
          type: ["integer", "null"]
        internalDeliveryNoteNumber:
          type: string
          minLength: 1
        internalDeliveryNoteDate:
          type: string
          format: date
        deliveryDestination:
          type: string
          enum: [administrative_building, health_facility, ambulance_point]
        notes:
          type: ["string", "null"]
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
...